Add timeouts
Added connect and read timeouts, set to 5s and 30s respectively. This means the connection should timeout if it fails to contact Telegram within 5s, or connects but doesn't hear back from Telegram (regarding the long-polling) within 30s (bit of leeway over the 20s polling wait).

This should help recover the application if the connection to Telegram is broken (instead of waiting forever and needing a reboot to fix).